chore: move server to subdirectory chore: specify dependencies for server (as optional) chore: add README
