+++ package.json
@@ -33,6 +33,10 @@
     "typescript": "^4.3.2",
     "ws": "^8.13.0"
   },
+  "optionalDependencies": {
+    "ws": "^8.13.0",
+    "@types/ws": "^8.5.4"
+  },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com/"
   },
