+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from "./client";
 export * from "./generateRandomId";
 export * from "./frame";
-export * from "./server";
 export * from "./session";
 export * from "./types";
